+++ test/test_helper.rb
@@ -12,5 +12,6 @@ require "trailblazer/operation/builder"
 require "trailblazer/operation/model"
 require "trailblazer/operation/contract"
 require "trailblazer/operation/representer"
+require "trailblazer/operation/present"
 
 # require "raise"